+++ tests/testthat/test-slide.R
@@ -13,7 +13,6 @@ test_that("Invalid input", {
   expect_error(slider(x, -6), "larger")
   expect_error(slider(list()), "larger")
   expect_error(slider(x, .bind = TRUE), "only accepts list")
-  expect_error(slider(list(x = 1.5, y = 1L), .bind = TRUE), "different types.")
 })
 
 test_that("check .align", {
@@ -74,6 +73,10 @@ test_that("slider() & pslider()", {
   expect_equal(
     unname(slider(lst, .size = 2, .partial = TRUE, .bind = TRUE, .fill = NA_integer_)[[1]]),
     c(NA, 1:5)
+  )
+  expect_equal(
+    slider(list(x = 1.5, y = 1L), .size = 2, .bind = TRUE),
+    list(c(1.5, 1.0))
   )
 })